+++ victorious/VictoriousIOSSDK/DeviceExperimentsRequest.swift
@@ -21,14 +21,14 @@ public struct DeviceExperimentsRequest: RequestType {
         return NSURLRequest(URL: NSURL(string: "/api/device/experiments")!)
     }
     
-    public func parseResponse(response: NSURLResponse, toRequest request: NSURLRequest, responseData: NSData, responseJSON: JSON) throws -> (experiments: [DeviceExperiment], defaultExperimentIDs: [Int64]) {
+    public func parseResponse(response: NSURLResponse, toRequest request: NSURLRequest, responseData: NSData, responseJSON: JSON) throws -> (experiments: [DeviceExperiment], defaultExperimentIDs: [Int]) {
             guard let experimentsJSON = responseJSON["payload"].array,
                 let defaultExperimentsJSON = responseJSON["experiment_ids"].array else {
                     throw ResponseParsingError()
             }
             
             let deviceExperiments = experimentsJSON.flatMap { DeviceExperiment(json: $0) }
-            let defaultExperiments = defaultExperimentsJSON.flatMap{Int64($0.stringValue)}
+            let defaultExperiments = defaultExperimentsJSON.flatMap{Int($0.stringValue)}
             return (deviceExperiments,defaultExperiments)
     }
 }
