+++ victorious/VictoriousIOSSDK/Models/User.swift
@@ -19,7 +19,7 @@ public enum ProfileStatus: String {
 
 /// A struct representing a user's information
 public struct User {
-    public let userID: Int64
+    public let userID: Int
     public let email: String?
     public let name: String?
     public let status: ProfileStatus?
@@ -29,24 +29,24 @@ public struct User {
     public let isCreator: Bool?
     public let isDirectMessagingDisabled: Bool?
     public let isFollowedByMainUser: Bool?
-    public let numberOfFollowers: Int64?
-    public let numberOfFollowing: Int64?
+    public let numberOfFollowers: Int?
+    public let numberOfFollowing: Int?
     public let profileImageURL: String?
     public let tokenUpdatedAt: NSDate?
     public let previewImageAssets: [ImageAsset]?
-    public let maxVideoUploadDuration: Int64?
+    public let maxVideoUploadDuration: Int?
 }
 
 extension User {
     public init?(json: JSON) {
-        let userIDFromJSON: Int64
+        let userIDFromJSON: Int
         let dateFormatter = NSDateFormatter( format: DateFormat.Standard )
         
         // Check for "id" as either a string or a number, because the back-end is inconsistent.
         if let userIDString = json["id"].string,
-           let userIDNumber = Int64(userIDString) {
+           let userIDNumber = Int(userIDString) {
             userIDFromJSON = userIDNumber
-        } else if let userIDValue = json["id"].int64 {
+        } else if let userIDValue = json["id"].int {
             userIDFromJSON = userIDValue
         } else {
             return nil
@@ -68,10 +68,10 @@ extension User {
         isCreator                   = json["isCreator"].bool ?? false
         isDirectMessagingDisabled   = json["is_direct_message_disabled"].bool
         isFollowedByMainUser        = json["am_following"].bool ?? false
-        numberOfFollowers           = Int64(json["number_of_followers"].stringValue)
-        numberOfFollowing           = Int64(json["number_of_following"].stringValue)
+        numberOfFollowers           = Int(json["number_of_followers"].stringValue)
+        numberOfFollowing           = Int(json["number_of_following"].stringValue)
         profileImageURL             = json["profile_image"].string
-        maxVideoUploadDuration      = Int64(json["max_video_duration"].stringValue)
+        maxVideoUploadDuration      = Int(json["max_video_duration"].stringValue)
         
         if let dateString = json["token_updated_at"].string {
             self.tokenUpdatedAt = dateFormatter.dateFromString(dateString)
